[agent] rename utils to internal
