+++ pkg/infra/teams/team1/internal/fit_data.go
@@ -6,7 +6,7 @@
 * This file or its contents can not be copied and/or used
 * without the express permission of Team 1, SOMAS2022
 *******************************************************/
-package utils
+package internal
 
 import (
 	"fmt"
