+++ pkg/infra/teams/team1/internal/q_values.go
@@ -1,4 +1,12 @@
-package utils
+/*******************************************************
+* Copyright (C) 2022 Team 1 @ SOMAS2022
+*
+* This file is part of SOMAS2022.
+*
+* This file or its contents can not be copied and/or used
+* without the express permission of Team 1, SOMAS2022
+*******************************************************/
+package internal
 
 import (
 	"infra/game/state"
