+++ pkg/infra/teams/team1/social_capital.go
@@ -4,7 +4,7 @@ import (
 	"infra/game/agent"
 	"infra/game/decision"
 	"infra/game/message"
-	"infra/teams/team1/utils"
+	"infra/teams/team1/internal"
 	"sort"
 )
 
@@ -47,15 +47,15 @@ func (s *SocialAgent) updateSocialCapital(self agent.BaseAgent, fightDecisions d
 	agentState := view.AgentState()
 
 	// Calculate how cooperative agents own action was
-	cooperativeQ := utils.CooperationQ(self.AgentState())
-	cooperationScale := utils.Normalise(cooperativeQ)
+	cooperativeQ := internal.CooperationQ(self.AgentState())
+	cooperationScale := internal.Normalise(cooperativeQ)
 	selfAction, _ := choices.Get(self.ID())
 	selfCooperation := cooperationScale[int(selfAction)]
 
 	// Update socialCapital values
 	for agentID := range s.socialCapital {
 		// Decay existing socialCapital values
-		s.socialCapital[agentID] = utils.DecayArray(s.socialCapital[agentID])
+		s.socialCapital[agentID] = internal.DecayArray(s.socialCapital[agentID])
 
 		// If agent did an action, update socialCapital based on action
 		action, exists := choices.Get(agentID)
@@ -64,10 +64,10 @@ func (s *SocialAgent) updateSocialCapital(self agent.BaseAgent, fightDecisions d
 			otherAgentState, _ := agentState.Get(agentID)
 
 			// Calculate how cooperative each action is in other agents current state
-			cooperativeQ := utils.HiddenCooperationQ(otherAgentState)
+			cooperativeQ := internal.HiddenCooperationQ(otherAgentState)
 
 			// Put actions on linear scale from -1 (least cooperative) to 1 (most cooperative)
-			cooperationScale := utils.Normalise(cooperativeQ)
+			cooperationScale := internal.Normalise(cooperativeQ)
 
 			// Calculate update of trustworthiness based on how cooperative action was
 			deltaTrust := 0.1 * cooperationScale[int(action)]
@@ -76,7 +76,7 @@ func (s *SocialAgent) updateSocialCapital(self agent.BaseAgent, fightDecisions d
 			deltaHonour := 0.1 * (cooperationScale[int(action)] - selfCooperation)
 
 			// Update the socialCapital array based on calculated delta for trustworthiness and honour
-			s.socialCapital[agentID] = utils.BoundArray(utils.AddArrays(s.socialCapital[agentID], [4]float64{0.0, 0.0, deltaTrust, deltaHonour}))
+			s.socialCapital[agentID] = internal.BoundArray(internal.AddArrays(s.socialCapital[agentID], [4]float64{0.0, 0.0, deltaTrust, deltaHonour}))
 		}
 	}
 }
@@ -159,7 +159,7 @@ func (s *SocialAgent) receiveGossip(m message.ArrayInfo, sender string) {
 	for _, about := range m.GetStringArr() {
 		sc := s.socialCapital[about]
 		sc[1] += sign * senderPerception * 0.1 * sc[1]
-		sc = utils.BoundArray(sc)
+		sc = internal.BoundArray(sc)
 		s.socialCapital[about] = sc
 	}
 }
